move nim to inference
