+++ plugins/flytekit-inference/flytekitplugins/inference/__init__.py
@@ -1,0 +1,3 @@
+from sidecar_template import ModelInferenceTemplate
+
+from .nim.serve import nim
